+++ clang/include/clang/Sema/Sema.h
@@ -1166,12 +1166,6 @@ public:
   void checkTypeSupport(QualType Ty, SourceLocation Loc,
                         ValueDecl *D = nullptr);
 
-  /// Check if the floating-point type is supported in the current language
-  /// mode and target. If \p DiagnoseTarget is true, diagnose offload code
-  /// according to whether its enclosing function is emitted.
-  bool checkFloatingPointTypeSupport(QualType Ty, SourceLocation Loc,
-                                     bool DiagnoseTarget = false);
-
   /// ImpCastExprToType - If Expr is not of type 'Type', insert an implicit
   /// cast.  If there is already an implicit cast, merge into the existing one.
   /// If isLvalue, the result of the cast is an lvalue.
@@ -2696,6 +2690,14 @@ public:
   ExprResult ConvertVectorExpr(Expr *E, TypeSourceInfo *TInfo,
                                SourceLocation BuiltinLoc,
                                SourceLocation RParenLoc);
+
+  /// ConvertFromArbitraryFPExpr - Handle
+  /// __builtin_elementwise_convert_from_<format>
+  ExprResult
+  ConvertFromArbitraryFPExpr(Expr *E,
+                             ConvertFromArbitraryFPExpr::ArbitraryFPFormat Fmt,
+                             TypeSourceInfo *TInfo, SourceLocation BuiltinLoc,
+                             SourceLocation RParenLoc);
 
   static StringRef GetFormatStringTypeName(FormatStringType FST);
   static FormatStringType GetFormatStringType(StringRef FormatFlavor);
@@ -7740,6 +7742,16 @@ public:
   ExprResult ActOnConvertVectorExpr(Expr *E, ParsedType ParsedDestTy,
                                     SourceLocation BuiltinLoc,
                                     SourceLocation RParenLoc);
+
+  /// ActOnConvertFromArbitraryFPExpr - create a new convert-from-arbitrary-fp
+  /// expression from the provided arguments.
+  ///
+  /// __builtin_elementwise_convert_from_<format>( value, dst type )
+  ///
+  ExprResult ActOnConvertFromArbitraryFPExpr(
+      Expr *E, ConvertFromArbitraryFPExpr::ArbitraryFPFormat Fmt,
+      ParsedType ParsedDestTy, SourceLocation BuiltinLoc,
+      SourceLocation RParenLoc);
 
   //===---------------------------- OpenCL Features -----------------------===//