+++ clang/lib/AST/StmtPrinter.cpp
@@ -1995,6 +1995,15 @@ void StmtPrinter::VisitConvertVectorExpr(ConvertVectorExpr *Node) {
   OS << ")";
 }
 
+void StmtPrinter::VisitConvertFromArbitraryFPExpr(
+    ConvertFromArbitraryFPExpr *Node) {
+  OS << Node->getBuiltinName() << "(";
+  PrintExpr(Node->getSrcExpr());
+  OS << ", ";
+  Node->getTypeSourceInfo()->getType().print(OS, Policy);
+  OS << ")";
+}
+
 void StmtPrinter::VisitInitListExpr(InitListExpr* Node) {
   if (Node->getSyntacticForm()) {
     Visit(Node->getSyntacticForm());