+++ clang/test/CodeGenCXX/mangle-exprs.cpp
@@ -399,3 +399,21 @@ namespace type_trait {
   // CHECK-LABEL: define {{.*}} @_ZN10type_trait1gIiEEvDTplu23__is_trivially_copyableiEcvT__EE
   template void g<int>(int);
 }
+
+namespace convert_from_arbitrary_fp {
+  template <typename T>
+  auto e5m2(T value) ->
+      decltype(__builtin_elementwise_convert_from_f8e5m2(value, float)) {
+    return {};
+  }
+  // CHECK-LABEL: define weak_odr noundef float @_ZN25convert_from_arbitrary_fp4e5m2IhEEDTu41__builtin_elementwise_convert_from_f8e5m2Xfp_EfEET_(
+  template float e5m2<unsigned char>(unsigned char);
+
+  template <typename T>
+  auto e4m3fn(T value) ->
+      decltype(__builtin_elementwise_convert_from_f8e4m3fn(value, float)) {
+    return {};
+  }
+  // CHECK-LABEL: define weak_odr noundef float @_ZN25convert_from_arbitrary_fp6e4m3fnIhEEDTu43__builtin_elementwise_convert_from_f8e4m3fnXfp_EfEET_(
+  template float e4m3fn<unsigned char>(unsigned char);
+}